backend/requirements.txt
acres==0.3.0
aiofiles==24.1.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.1
etelemetry==0.3.1
fastapi==0.115.8
filelock==3.17.0
fitz==0.0.1.dev2
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
iniconfig==2.0.0
isodate==0.6.1
looseversion==1.3.0
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pillow==11.1.0
pluggy==1.5.0
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tokenizers==0.21.0
tqdm==4.67.1
traits==7.0.2
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
XlsxWriter==3.2.2
